+++ cms/patients/routes.py
@@ -1,4 +1,5 @@
 from flask import Blueprint, render_template, redirect, url_for, request, jsonify
+from flask_login import current_user
 from datetime import datetime
 from cms import db
 from cms.models import Patient, PatientSchema
@@ -8,6 +9,8 @@ patients = Blueprint('patients', __name__)
 
 @patients.route('/', methods=['GET'])
 def index():
+    if current_user.position == 'patient':
+        return redirect(url_for('front_page.index'))
     patients = Patient.query.all()
     patients_schema = PatientSchema(many=True)
     output = patients_schema.dump(patients).data
@@ -15,14 +18,18 @@ def index():
 
 @patients.route('/<patient>', methods=['GET'])
 def view(patient):
+    if current_user.position == 'patient':
+        return redirect(url_for('front_page.index'))
     patient = Patient.query.get(patient)
     patient_schema = PatientSchema()
     output = patient_schema.dump(patient).data
-    # return jsonify(output)
+
     return render_template('patients/view.html', patient=patient)
 
 @patients.route('/create', methods=['GET'])
 def create():
+    if current_user.position == 'patient':
+        return redirect(url_for('front_page.index'))
     form = CreatePatientForm()
     return render_template('patients/create.html', form=form)
 
@@ -43,6 +50,8 @@ def save():
 
 @patients.route('/<patient>/edit', methods=['GET'])
 def edit(patient):
+    if current_user.position == 'patient':
+        return redirect(url_for('front_page.index'))
     patient = Patient.query.get(patient)
     form = EditPatientForm(obj = patient)
     print(form.date_of_birth.data)
@@ -66,7 +75,10 @@ def update():
 
 @patients.route('<patient>/delete/', methods=['GET'])
 def delete(patient):
+    if current_user.position == 'patient':
+        return redirect(url_for('front_page.index'))
     patient = Patient.query.get(int(patient))
     db.session.delete(patient)
     db.session.commit()
     return redirect(url_for('patients.index'))
+
